api/requirements.txt
azure-storage-blob==12.19.0
azure-ai-formrecognizer==3.3.2
azure-ai-textanalytics==5.3.0
azure-cognitiveservices-speech==1.34.0
azure-search-documents==11.4.0b8
azure-storage-blob==12.19.0
azure-identity==1.15.0
fastapi==0.108.0
google==3.0.0
langchain==0.1.0
langchain-community==0.0.16
langchain-openai==0.0.2
openai==1.7.1
pydantic==2.9.1
pandas
requests
scipy==1.12.0
snowflake-connector-python==3.6.0
snowflake-sqlalchemy==1.5.1
tiktoken==0.5.2
uvicorn==0.25.0
python-dotenv
numexpr
instructor==1.3.7
